[DX12][common] Change loop counter type to size_t
It's a size, duh. Also change the type of `size` to match.

+++ DirectX/common/utils/keyUtils.cpp
@@ -37,7 +37,7 @@ ConfigKeySet::ConfigKeySet(const std::string& keys) {
 
 std::string parseConfigKeys(const std::string& keys) {
   std::string result;
-  for (unsigned i = 0; i < keys.size(); ++i) {
+  for (size_t i = 0; i < keys.size(); ++i) {
     char c = keys[i];
     if (c == 'S' || c == 'E') {
       const char* begin = &keys[++i];
